requirement.txt
absl-py==0.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bio==1.6.2
biopython==1.83
biothings-client==0.3.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clang==5.0
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
filelock==3.13.1
flatbuffers==1.12
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprofiler-official==1.0.0
grpcio==1.63.0
h5py==3.1.0
hiddenlayer==0.3
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.4
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
logomaker==0.8.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.3.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mygene==3.2.2
narwhals==1.34.0
natsort==8.4.0
networkx==3.0
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==3.20.0
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.11.2
six==1.15.0
smmap==5.0.2
streamlit==1.40.1
sympy==1.12
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.1.0+cu121
torch-geometric==2.6.1
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.1
watchdog==4.0.2
Werkzeug==3.0.3
wrapt==1.12.1
xgboost==2.1.1
xlrd==2.0.1
yarl==1.14.0
zipp==3.18.2
